Simplify HTTP dispatcher; remove ErrorHandlerInterface
