--- src/Http/ErrorHandlerInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Spiral\RoadRunnerBridge\Http;
-
-interface ErrorHandlerInterface
-{
-    public function handle(\Throwable $e): void;
-}
